requirements.txt
langchain
wikipedia
ddgs
langchain-community
langchain-openai
langchain-google-genai
langchain-anthropic
python-dotenv
pydantic
google-generativeai